requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
dbt-postgres>=1.7.0  # Allow newer versions that may have Python 3.12 fixes
pandas==2.1.4
numpy>=1.26.0
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.0.219
dagster==1.5.10
dagster-webserver==1.5.10